producer/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
kafka-python==2.0.2
python-dateutil==2.8.2
requests==2.25.1
six==1.15.0
sseclient==0.0.27
urllib3==1.26.5
